In rte_config.h change RTE_ETHDEV_QUEUE_STAT_CNTRS to 128
This define specifies the maximum number of queues for which statistics will be
received from the network card. If 2 numa nodes are used and n jobs are running
on each node, then n queues will be created for RX and 2n+1 queues for TX.
Therefore, in order not to lose statistics on queues, it is necessary that the
inequality 2n + 1 <= RTE_ETHDEV_QUEUE_STAT_CNTRS be fulfilled.